Remove excessive code from course provisioning routines.
For course context, provision_course_internal.php is used when course has
not been provisioned yet, so there is no way to retrieve server/key from
foldermap record as they simply not there yet. In fact those removed
statements are never reached in reality.

For system context, provision_course.php is used irrespective whether
course has been provisioned, again all requred server/key data is in the
config, no need to check if it has been stored in foldermap. And again, in
reality those statements are never reached because although
provision_course.php has some coverage for the case when being called from
the course level (in which case $selectedserver is not defined and supposed
to be retrieved from the foldermap table), this is never being used as from
the course context provision_course_internal.php is being used instead.