Add missing upgrade to move panopto server settings to new variables.
This ideally should be a part of 7741c05fc55 that introduce multiple
servers. The server host/key remain undefined after upgrade since new
variable names are used. The patch will only be useful for those upgrading
from earlier version than 2014121702, it is generally a bad practice to
modify upgrade.php retrospectively, but important in this case to fix
panopto connection for the small proportion of users who has not upgraded
yet.